feat: dashboard blueprint, api & json documentation
