+++ docs/JSON.md
@@ -2,6 +2,9 @@
 
 ### General response:
 - status: boolean true or false
+- statusText: message text as "accepted" or "rejected"
+- data: received data or "none"
+- error: potential error text or "none"
 
 Example:
 ```JSON
@@ -15,6 +18,11 @@ Example:
 
 ### Dashboard response:
 - status: boolean true or false
+- statusText: message text as "accepted" or "rejected"
+- data: received data or "none"
+- currency: currency type "RON", "EUR" or "USD"
+- list: list by day
+- error: potential error text or "none"
 
 Example:
 ```JSON
