+++ server/__init__.py
@@ -2,5 +2,5 @@ from .settings import get_db_info
 from .manager import PostgresqlManager, CurrencyManager
 from .enum import Category, Currency
 from .model import Report
-from .blueprints import import_api
+from .blueprints import import_api, dashboard_api
