+++ server/blueprints/__init__.py
@@ -1,1 +1,2 @@
 from .import_api import import_api
+from .dashboard_api import dashboard_api
